fix: resovle comment feedback
