+++ apps/app/src/App.tsx
@@ -5,7 +5,7 @@
  * @format
  */
 
-import React from 'react'
+import React, { useEffect } from 'react'
 import { useColorScheme } from 'react-native'
 import { NavigationContainer } from '@react-navigation/native'
 import { createNativeStackNavigator } from '@react-navigation/native-stack'
@@ -14,6 +14,10 @@ import ErrorBoundary from 'react-native-error-boundary'
 
 import { LoginScreen, RegisterScreen, WebviewScreen, ErrorScreen } from '@/pages'
 import { RootStackParamList } from '@/types'
+import PushNotificationService from '@/services/pushNotificationService.ts'
+import { getToken } from '@/stores'
+import { IRequestNotificationToken } from '@fienmee/types'
+import { submitFCMToken } from '@/api'
 
 function App(): React.JSX.Element {
     const isDarkMode = useColorScheme() === 'dark'
@@ -22,6 +26,24 @@ function App(): React.JSX.Element {
         backgroundColor: isDarkMode ? Colors.darker : Colors.lighter,
     }
     const Stack = createNativeStackNavigator<RootStackParamList>()
+
+    useEffect(() => {
+        PushNotificationService.setMessageHandler()
+        PushNotificationService.listenRefreshToken(async token => {
+            const credential = await getToken()
+            const info = await PushNotificationService.getDeviceInfo()
+            const request: IRequestNotificationToken = {
+                body: {
+                    token: token,
+                    deviceId: info.deviceId,
+                    platform: info.platform,
+                },
+                accessToken: credential.accessToken,
+            }
+            console.log(request)
+            await submitFCMToken(request)
+        })
+    }, [])
 
     return (
         <ErrorBoundary>
